separate out rpc to sync and async apis (#26570)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26570

This diff splits out the async and sync rpc implementations to their
own functions. This way, we can have stronger typehinting and make the type of
RPC being done more explicit to our users, as opposed to them having to pass
in an async flag to the functiohn.
ghstack-source-id: 91023006

Test Plan: Tests in test/test_rpc.py should all pass.

Differential Revision: D17509975

fbshipit-source-id: cc331ec1fdd50c31fa1fae3ba73ddb95d17dad64