Move imports to after is_available in test_rpc to fix windows builds (#27126)
Summary:
These imports need to go after the check to not break windows since rpc is not supported on windows. I ran the linter and that moved it to before the check in https://github.com/pytorch/pytorch/issues/26570. This wasn't caught by the windows tests in the PR since it was failing due to a JIT-related reason: (https://ci.pytorch.org/jenkins/job/pytorch-builds/job/pytorch-win-ws2016-cuda9-cudnn7-py3-test2/49433/console). In the future, we should probably have a warning/a comment/some message about discouraging running linters? Thanks to peterjc123 for flagging this.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27126

Differential Revision: D17682761

Pulled By: rohan-varma

fbshipit-source-id: 300c74290d734eb8e5880104d2b76dd64217b696