Fix aten::format regex for clang8 (#28916)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28916

The previous regex caused a `std::regex_error` under clang8 complaining about `error_brack`, which is strange because the square brackets are balanced. Seems like a stdlib bug to me. So to workaround this, I've switched to the older regex with a non-greedy match in the inner atom

Test Plan: Imported from OSS

Differential Revision: D18232654

Pulled By: jamesr66a

fbshipit-source-id: f82a9a24acf090010b03f23454d2b0f7a1e3589e