Add functional overloads for pixelshuffle, pooling, upsampling, vision (#29359)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29359

This PR adds functional overloads that take the full set of arguments (instead of just Options) for the following functionals:
- pixelshuffle
- pooling
- upsampling
- vision

These new functionals lives in the `torch::nn::functional::detail` namespace and they are only meant to be called from the module forward methods (i.e. they are not public API). This is in preparation for the future change where we make module Options and functional Options two different classes, because if the module forward method has to construct a new functional Options object every time it runs it will be pretty silly and bad performance.

Test Plan: Imported from OSS

Differential Revision: D18376978

Pulled By: yf225

fbshipit-source-id: 4ea8d359e7efde0d741eff79faad6b24b2a5d804