Declare Dimname's kWildcard as extern instead of static (#29384)
Summary:
Fixes https://github.com/pytorch/pytorch/issues/27627

The variable being declared in a header as `static` meant that the global variable is initialized in every source file that includes it. This is particularly problematic when included in AVX source files as it generates SIGILL on older hardware.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29384

Differential Revision: D18380379

Pulled By: zou3519

fbshipit-source-id: 0dcd87db01c468a5c9ddb2c695528b85ed2e1504